Modify as per Roma's comments:  - rename files&folder  - minor Observations change - remove redundant words in template  - remove any type from saga
