+++ src/pages/comparison/components/rate.component.tsx
@@ -5,7 +5,6 @@ import clases from './rate-logic.service';
 import { RateProps } from '../interfaces/pet.interface';
 
 export const Rate: React.FC<RateProps> = props => {
-
   const items = clases(props.rate).map((el, index) => {
     return <div key={index} className={el}></div>;
   });
