+++ src/pages/pet/components/observations/observations.component.tsx
@@ -6,19 +6,19 @@ export const ObservationsComponent: React.FC<ObservationProps> = props => {
   const { country, lifetime, height, weight, price, carePrice, careTime } = props.data;
   const blocks = {
     Країна: country,
-    'Тривалість життя': `${lifetime} років`,
-    Висота: `${height} см`,
-    Вага: `${weight} кг`,
-    Вартість: `${price} грн`,
-    'Ціна утримання': `${carePrice} грн`,
+    'Тривалість життя': lifetime,
+    Висота: height,
+    Вага: weight,
+    Вартість: price,
+    'Ціна утримання': carePrice,
     'Час догляду': `${careTime} год на день`,
   };
   const blocksArray = Object.entries(blocks);
   return (
     <ul className={style.obsContainer}>
-      {blocksArray.map((entry, index) => (
+      {blocksArray.map(([title, content], index) => (
         <li className={style.obsItem} key={index}>
-          <p className={style.obsTitle}>{entry[0]}</p> <p className={style.plainText}>{entry[1]}</p>
+          <p className={style.obsTitle}>{title}</p> <p className={style.plainText}>{content}</p>
         </li>
       ))}
     </ul>
