--- src/store/pet/action/pet.action.ts
@@ -1,5 +1,0 @@
-import { createAsyncAction } from 'typesafe-actions';
-import {PetProfile} from '../../../shared/models';
-
-export const fetchPetProfile = createAsyncAction('FETCH_REQUEST','FETCH_SUCCESS','FETCH_ERROR'
-)<number, PetProfile, string>();
